+++ Model/src/main/java/org/gusdb/wdk/model/user/dataset/event/UserDatasetEvent.java
@@ -7,11 +7,11 @@ import org.gusdb.wdk.model.user.dataset.UserDatasetType;
 
 public class UserDatasetEvent {
   private Set<String> projectsFilter; // null if no filter
-  private Integer userDatasetId;
+  private Long userDatasetId;
   private UserDatasetType userDatasetType;
   private Long eventId;
   
-  public UserDatasetEvent(Long eventId, Set<String> projectsFilter, Integer userDatasetId, UserDatasetType userDatasetType) {
+  public UserDatasetEvent(Long eventId, Set<String> projectsFilter, Long userDatasetId, UserDatasetType userDatasetType) {
     this.eventId = eventId;
     this.projectsFilter = projectsFilter;
     this.userDatasetId = userDatasetId;
@@ -19,7 +19,7 @@ public class UserDatasetEvent {
   }
   
   public Set<String> getProjectsFilter() { return Collections.unmodifiableSet(projectsFilter); }
-  public Integer getUserDatasetId() { return userDatasetId; }
+  public Long getUserDatasetId() { return userDatasetId; }
   public UserDatasetType getUserDatasetType() { return userDatasetType; }
   public Long getEventId() { return eventId; }
 }
