+++ Model/src/main/java/org/gusdb/wdk/model/user/dataset/event/UserDatasetInstallEvent.java
@@ -7,15 +7,17 @@ import org.gusdb.wdk.model.user.dataset.UserDatasetDependency;
 import org.gusdb.wdk.model.user.dataset.UserDatasetType;
 
 public class UserDatasetInstallEvent extends UserDatasetEvent {
+
   private Set<UserDatasetDependency> dependencies;
-  private Integer ownerUserId;
-  
-  public UserDatasetInstallEvent(Long eventId, Set<String> projectsFilter, Integer userDatasetId, UserDatasetType userDatasetType, Integer ownerUserId, Set<UserDatasetDependency> dependencies) {
+  private Long ownerUserId;
+
+  public UserDatasetInstallEvent(Long eventId, Set<String> projectsFilter, Long userDatasetId,
+      UserDatasetType userDatasetType, Long ownerUserId, Set<UserDatasetDependency> dependencies) {
     super(eventId, projectsFilter, userDatasetId, userDatasetType);
     this.dependencies = dependencies;
     this.ownerUserId = ownerUserId;
   }
 
   public Set<UserDatasetDependency> getDependencies() { return Collections.unmodifiableSet(dependencies); }
-  public Integer getOwnerUserId() { return ownerUserId; }
+  public Long getOwnerUserId() { return ownerUserId; }
 }
